moved coloring to colors.js
