+++ src/components/home/queryresults/value/Value.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react";
 import useWebSocket from 'react-use-websocket';
 import api from '../../../../api';
+import colors from '../../../../colors';
 import PropTypes from "prop-types";
 
 const propTypes = {
@@ -74,22 +75,22 @@ function Value(props) {
         if (pvSeverity !== undefined) {
             if (pvSeverity === "INVALID") {
                 return (
-                    <div style={{ color: api.SEV_COLORS["INVALID"] }}>{pvValue} ({pvSeverity})</div>
+                    <div style={{ color: colors.SEV_COLORS["INVALID"] }}>{pvValue} ({pvSeverity})</div>
                 );
             }
             else if (pvSeverity === "UNDEFINED") {
                 return (
-                    <div style={{ color: api.SEV_COLORS["UNDEFINED"] }}>{pvSeverity}</div>
+                    <div style={{ color: colors.SEV_COLORS["UNDEFINED"] }}>{pvSeverity}</div>
                 );
             }
             else if (pvSeverity === "NONE") {
-                textColor = api.SEV_COLORS["OK"];
+                textColor = colors.SEV_COLORS["OK"];
             }
             else if (pvSeverity === "MINOR") {
-                textColor = api.SEV_COLORS["MINOR"];
+                textColor = colors.SEV_COLORS["MINOR"];
             }
             else if (pvSeverity === "MAJOR") {
-                textColor = api.SEV_COLORS["MAJOR"];
+                textColor = colors.SEV_COLORS["MAJOR"];
             }
         }
         if (pvUnit !== undefined) {
